requirements.txt
python-telegram-bot==13.11
psycopg2-binary
pandas
pyzbar
openpyxl
xlsxwriter
